+++ Core/Config.h
@@ -321,7 +321,6 @@ public:
 	float fUITint;
 	float fUISaturation;
 
-	bool bTextureBackoffCache;
 	bool bVertexDecoderJit;
 	int iAppSwitchMode;
 	bool bFullScreen;