+++ GPU/Vulkan/ShaderManagerVulkan.cpp
@@ -118,10 +118,9 @@ VulkanFragmentShader::~VulkanFragmentShader() {
 		if (shaderModule) {
 			vulkan_->Delete().QueueDeleteShaderModule(shaderModule);
 		}
-		vulkan_->Delete().QueueCallback([](VulkanContext *vulkan, void *m) {
-			auto module = (Promise<VkShaderModule> *)m;
+		vulkan_->Delete().QueueCallback([module = module_](VulkanContext *vulkan) {
 			delete module;
-		}, module_);
+		});
 	}
 }
 
@@ -150,10 +149,9 @@ VulkanVertexShader::~VulkanVertexShader() {
 		if (shaderModule) {
 			vulkan_->Delete().QueueDeleteShaderModule(shaderModule);
 		}
-		vulkan_->Delete().QueueCallback([](VulkanContext *vulkan, void *m) {
-			auto module = (Promise<VkShaderModule> *)m;
+		vulkan_->Delete().QueueCallback([module = module_](VulkanContext *vulkan) {
 			delete module;
-		}, module_);
+		});
 	}
 }
 
@@ -182,10 +180,9 @@ VulkanGeometryShader::~VulkanGeometryShader() {
 		if (shaderModule) {
 			vulkan_->Delete().QueueDeleteShaderModule(shaderModule);
 		}
-		vulkan_->Delete().QueueCallback([](VulkanContext *vulkan, void *m) {
-			auto module = (Promise<VkShaderModule> *)m;
+		vulkan_->Delete().QueueCallback([module = module_](VulkanContext *vulkan) {
 			delete module;
-		}, module_);
+		});
 	}
 }