+++ libretro/libretro.cpp
@@ -797,16 +797,6 @@ static void check_variables(CoreParameter &coreParam)
       else
          g_Config.iSkipGPUReadbackMode = (int)SkipGPUReadbackMode::SKIP;
    }
-
-   var.key = "ppsspp_lazy_texture_caching";
-   if (environ_cb(RETRO_ENVIRONMENT_GET_VARIABLE, &var) && var.value)
-   {
-      if (!strcmp(var.value, "disabled"))
-         g_Config.bTextureBackoffCache = false;
-      else
-         g_Config.bTextureBackoffCache = true;
-   }
-
    var.key = "ppsspp_spline_quality";
    if (environ_cb(RETRO_ENVIRONMENT_GET_VARIABLE, &var) && var.value)
    {